Making the entire project data oriented.
