+++ src/main/java/us/jmay/Menu.java
@@ -1,0 +1,6 @@
+package us.jmay;
+
+public class Menu {
+    public String section;
+    public MenuDisplay[] menuDisplays;
+}
