+++ src/main/java/us/jmay/MenuItem.java
@@ -1,0 +1,10 @@
+package us.jmay;
+
+public class MenuItem {
+    public String name;
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
