+++ src/main/java/us/jmay/utils/DataUtil.java
@@ -1,0 +1,12 @@
+package us.jmay.utils;
+
+import java.io.File;
+import java.net.URL;
+
+public class DataUtil {
+
+    public URL getLocationFile()
+    {
+        return this.getClass().getResource("locations.json");
+    }
+}
