+++ src/main/resources/locations.json
@@ -1,0 +1,10 @@
+{
+  "url": "https://dining.iastate.edu/wp-json/dining/menu-hours/get-single-location/?slug=%s&time=",
+
+  "locations": [
+    "seasons-marketplace-2-2",
+    "conversations-2",
+    "union-drive-marketplace-2-2",
+    "friley-windows-2-2"
+  ]
+}
