--- src/us/jmay/Menu.java
@@ -1,6 +1,0 @@
-package us.jmay;
-
-public class Menu {
-    public String section;
-    public MenuDisplay[] menuDisplays;
-}
