--- src/us/jmay/MenuItem.java
@@ -1,10 +1,0 @@
-package us.jmay;
-
-public class MenuItem {
-    public String name;
-
-    @Override
-    public String toString() {
-        return name;
-    }
-}
